+++ CHANGELOG.md
@@ -10,7 +10,13 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.4.0] - 2021-07-22
+
+- [Performance improvements.](https://github.com/taiki-e/parse-changelog/pull/8)
+
 ### Library
+
+- [Change `Release::notes` field from `String` to `&str`.](https://github.com/taiki-e/parse-changelog/pull/8)
 
 - [Add `parse_iter` for partial parsing changelog.](https://github.com/taiki-e/parse-changelog/pull/9)
 
@@ -62,7 +68,8 @@ No public API changes from 0.2.1.
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/parse-changelog/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/taiki-e/parse-changelog/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/taiki-e/parse-changelog/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/taiki-e/parse-changelog/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/taiki-e/parse-changelog/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/taiki-e/parse-changelog/compare/v0.2.0...v0.2.1